pyproject.toml
[project]
name = "portfolio-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6.3",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.3",
    "langchain-core>=0.3.58",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.1.2",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
